[MNT] Move _IEBlock to layers module
