+++ pytorch_forecasting/layers/_blocks/__init__.py
@@ -1,3 +1,4 @@
+from pytorch_forecasting.layers._blocks._ie_block import _IEBlock
 from pytorch_forecasting.layers._blocks._residual_block_dsipts import ResidualBlock
 
-__all__ = ["ResidualBlock"]
+__all__ = ["ResidualBlock", "_IEBlock"]
